conversion-system/voice/ai_agent/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Pipecat Framework (Voice AI Pipeline)
pipecat-ai==0.0.40

# LLM Providers
openai==1.12.0

# STT/TTS Providers
deepgram-sdk==3.2.7

# Audio Processing
numpy==1.26.3
scipy==1.12.0

# WebSocket Support
websockets==12.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Logging
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
